requirements.txt
Django==1.6.1
Pillow==2.2.2
South==0.7.6
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.13
boto==2.6.0
celery==3.1.7
dj-database-url==0.2.0
dj-static==0.0.5
django-celery==3.1.1
django-crispy-forms==1.4.0
django-lockdown==1.0
django-social-auth==0.7.28
django-storages==1.1.5
django-vanilla-views==1.0.2
httplib2==0.8
jsonfield==0.9.20
kombu==3.0.8
oauth2==1.5.211
psycopg2==2.4.5
python-openid==2.2.5
pytz==2013.8
requests==1.1.0
wsgiref==0.1.2
django-mailgun==0.2.2
honcho==0.5.0
django-pipeline==1.3.19
sorl-thumbnail==11.12
pylibmc==1.2.3
django-heroku-memcacheify==0.4
django-pagination==1.0.7
django-registration==1.0
django-settings-context-processor==0.2
simplejson==3.3.1
jsmin==2.0.8
stripe==1.11.0
mutagen==1.22
django-mail-templated==0.2.0
django-secure==1.0
django-endless-pagination==2.0
djangular==0.2.5
django-avatar==2.0
django-extensions==1.3.9
raven==5.1.1
